requirements.txt
#asgiref==3.3.1
#astroid==2.4.2
#cycler==0.10.0
Django==3.1.5
#isort==5.7.0
#kiwisolver==1.3.1
#lazy-object-proxy==1.4.3
matplotlib==3.3.3
#mccabe==0.6.1
numpy==1.19.5
Pillow==8.1.0
psycopg2-binary
#pylint==2.6.0
pynmea2==1.15.0
#pyparsing==2.4.7
#python-dateutil==2.8.1
#pytz==2020.5
scipy==1.6.0
#six==1.15.0
#sqlparse==0.4.1
#toml==0.10.2
#wrapt==1.12.1
django-language-fields == 1.0.0
djangorestframework == 3.12.2
